// src/api/form.types.ts
export type CallReplicationType = {
  tokenA_balance: string
  amount: string
  fee: string
  perDay: string
  strike: string
  expiration: string
  riskFree: string
  sigma: string
}

export type PutReplicationType = {
  tokenB_balance: string
  amount: string
  fee: string
  perDay: string
  strike: string
  expiration: string
  riskFree: string
  sigma: string
}